Add fast blend modes ZERO and ADD. Add TexEnvfv for rasterizer. Fix viewport offsets.
git-svn-id: http://bricks-os.googlecode.com/svn/trunk@781 9b99efca-2682-11de-ba08-0b5a79c0c79e

+++ gl/textures.cpp
@@ -411,8 +411,8 @@ CTexture::getTexel(TColor<int32_t> & c, float u, float v, float lod)
   int iU, iV;
   uint32_t texel;
 
-  iU = u * width;
-  iV = v * height;
+  iU = (int)(u * width);
+  iV = (int)(v * height);
   // Wrap
   iU &= iWidthMask_;
   iV &= iHeightMask_;
